refactor(renderer): extract finding_enrichment phase 3 — renderer extraction roadmap COMPLETE
The HIGH-coupling subset that v0.12.16 documented as "deferred until a
proper coupling-graph design" and v0.12.19 further deferred ("needs a
sibling source_resolution.py extraction first") now ships, completing
the v0.10.0 renderer extraction roadmap.

The path that unblocked this: v0.12.23's source_resolution.py
extraction. Before that, the 5 phase-3 functions depended on 6
helpers (_action_dotted_entry, _skip_decorators_to_def,
_resolve_dotted_to_code, _action_entry_callsite,
_resolve_frame_key_to_callsite, _bench_relative_display) that lived
in _internal.py. v0.12.23 moved the 6 helpers first as a clean
independent extraction; today's PR moves the remaining 5 + 3 sibling
helpers as the final phase.

Moved
-----
8 functions + 1 constant from _internal.py → finding_enrichment.py:
* _find_call_line_in_function_body — AST-walker for finding a
  callee's call line inside a parent function.
* _retarget_phase1_callsites_to_drilldown_leaf — re-aim phase-1
  finding callsites at the call site of the deepest user-code frame
  in their drill-down chain.
* _finding_to_dict (~200 LOC) — the main render-dict builder.
* _attach_representative_callsites — attach a representative
  callsite to SQL red-flag findings.
* _markdown_to_safe_html — render Markdown → sanitised HTML.
* _read_function_body_snippet — read a whole function body starting
  from its def line.
* _expand_self_time_snippets — narrow self-time hot-path snippets
  to the function signature line.
* _SQL_REDFLAG_FINDING_TYPES constant.

Out of optimus/renderer/_internal.py (now ~2,265 LOC; was ~2,928),
into existing optimus/renderer/finding_enrichment.py (extended from
~375 LOC → ~995 LOC).

Implementation
--------------
* Lazy imports of sibling submodules inside each function body
  rather than at module-top. Same pattern as phase 2 — keeps the
  submodule free of circular import risk.
* Standard back-import block at the top of _internal.py re-imports
  all 8 new names + the constant so call sites resolve unchanged.
* Structural-snapshot canary stays byte-identical — no fixture
  regeneration needed.
* Public-API contract preserved — _finding_to_dict,
  _markdown_to_safe_html, _read_function_body_snippet all still
  resolve via renderer.X through the package __init__.py dir-walk.

Renderer extraction is now complete
-----------------------------------
9 submodules total:
* source.py + syntax.py + time_format.py + visualization.py (v0.10.0)
* call_tree_renderer.py (v0.12.8)
* doc_event_renderer.py (v0.12.10)
* line_drilldown.py (v0.12.12)
* finding_enrichment.py (v0.12.16 + v0.12.19 + v0.12.26)
* source_resolution.py (v0.12.23)

The only thing left in _internal.py is the render() orchestrator
itself + its immediate inline helpers. The README's "keep integrated"
guidance for that final 800-LOC function still holds.

Unchanged
---------
* Behaviour identical — every call site resolves the same function
  through the back-import shim.
* analyze.py callers of renderer._finding_to_dict and
  renderer._markdown_to_safe_html stay green without changes.

Compatibility
-------------
No behaviour change. Pure refactor. Unit suite stays at 1870.

Verification
------------
* env/bin/python -m pytest optimus/tests/test_renderer_structure_snapshot.py -v
    → 14 passed in 0.24s (fingerprint matches, no fixture update)
* env/bin/python -m pytest optimus/tests/ -q
  --ignore=optimus/tests/test_pdf_export.py
    → 1870 passed, 1 skipped in 6.47s
* ruff check on touched files → All checks passed!
