+++ optimus/renderer/_internal.py
@@ -132,10 +132,18 @@ from optimus.renderer.doc_event_renderer import (
 # focused future PR with proper coupling-graph design.
 from optimus.renderer.finding_enrichment import (
 	_GROUPING_SEVERITY_RANK,
+	_SQL_REDFLAG_FINDING_TYPES,
 	_attach_drilldown_chains,
+	_attach_representative_callsites,
+	_expand_self_time_snippets,
+	_find_call_line_in_function_body,
 	_find_node_in_tree,
+	_finding_to_dict,
 	_group_findings_by_root_cause,
+	_markdown_to_safe_html,
 	_normalize_callsite,
+	_read_function_body_snippet,
+	_retarget_phase1_callsites_to_drilldown_leaf,
 	_root_cause_key,
 	_walk_drilldown_chain,
 )
@@ -1059,276 +1067,10 @@ def _e(text: object) -> str:
 # this module.
 
 
-def _find_call_line_in_function_body(
-	parent_filename: str,
-	parent_def_lineno: int,
-	callee_function: str,
-	*,
-	file_cache: dict | None = None,
-) -> int | None:
-	"""Return the lineno of the first call to ``callee_function`` inside
-	the function whose ``def`` begins at ``parent_def_lineno`` in
-	``parent_filename``. ``None`` if the source can't be read or no call
-	is found.
-
-	AST primary — locates the matching ``FunctionDef`` / ``AsyncFunctionDef``
-	node by name/lineno, then walks its body for ``Call`` expressions
-	whose target resolves to ``callee_function`` (matches both bare
-	``callee_function(...)`` via ``Name`` and ``obj.callee_function(...)``
-	via ``Attribute``).
-
-	Regex fallback when AST parse fails (truncated file, syntax error
-	elsewhere) — scans lines below the def for ``\\b<callee>\\s*\\(``
-	stopping at a same-or-lower indented ``def `` / ``class `` /
-	``async def `` or after 200 lines.
-
-	Reused across multiple findings in the same file via ``file_cache``
-	(the same per-render cache passed to ``_read_source_snippet``).
-	"""
-	if not parent_filename or not callee_function or not parent_def_lineno:
-		return None
-
-	# Read source through the shared cache so files visited by
-	# _read_source_snippet aren't re-read here.
-	if file_cache is not None and parent_filename in file_cache:
-		lines = file_cache[parent_filename]
-	else:
-		resolved = _resolve_source_path(parent_filename)
-		try:
-			with open(resolved, encoding="utf-8") as fh:
-				lines = fh.read().splitlines()
-		except Exception:
-			lines = None
-		if file_cache is not None:
-			file_cache[parent_filename] = lines
-
-	if not lines:
-		return None
-	source = "\n".join(lines)
-
-	# Strategy 1: AST parse.
-	try:
-		import ast as _ast
-		tree = _ast.parse(source, filename=parent_filename)
-	except Exception:
-		tree = None
-
-	if tree is not None:
-		# Find the function def closest to parent_def_lineno (defensive
-		# against the parent being a nested function — pick by name match
-		# AND minimum lineno distance).
-		candidates: list[_ast.AST] = []
-		for node in _ast.walk(tree):
-			if isinstance(node, (_ast.FunctionDef, _ast.AsyncFunctionDef)):
-				candidates.append(node)
-		if candidates:
-			# Filter by lineno proximity to the recorded def lineno (within
-			# a few lines is usually enough — pyinstrument's lineno can
-			# be the def or the first executed line of the body).
-			best = min(
-				candidates,
-				key=lambda n: abs((n.lineno or 0) - parent_def_lineno),
-			)
-			# Walk only inside the chosen function; collect Call linenos
-			# whose func name matches.
-			matches: list[int] = []
-			for sub in _ast.walk(best):
-				if not isinstance(sub, _ast.Call):
-					continue
-				func = sub.func
-				name = None
-				if isinstance(func, _ast.Name):
-					name = func.id
-				elif isinstance(func, _ast.Attribute):
-					name = func.attr
-				if name == callee_function and sub.lineno:
-					matches.append(sub.lineno)
-			if matches:
-				return min(matches)
-
-	# Strategy 2: regex fallback. Scan from parent_def_lineno+1 until
-	# a sibling def/class at the parent's indentation level or 200
-	# lines, whichever comes first. Stop at first match.
-	import re as _re
-	call_re = _re.compile(r"\b" + _re.escape(callee_function) + r"\s*\(")
-	# Parent's indentation level (number of leading spaces / tabs).
-	if not (1 <= parent_def_lineno <= len(lines)):
-		return None
-	parent_line = lines[parent_def_lineno - 1]
-	parent_indent = len(parent_line) - len(parent_line.lstrip())
-	# Body must be MORE indented than the parent def.
-	sentinel_re = _re.compile(
-		r"^(?P<indent>\s*)(?:async\s+def|def|class)\s",
-	)
-	max_scan = min(len(lines), parent_def_lineno + 200)
-	for i in range(parent_def_lineno, max_scan):
-		raw = lines[i]
-		# Strip trailing # comments naively (string-aware comment
-		# stripping is overkill here — the worst case is a false
-		# negative which falls back to the def-line behavior).
-		hash_idx = raw.find("#")
-		body = raw if hash_idx < 0 else raw[:hash_idx]
-		stripped = body.lstrip()
-		if not stripped:
-			continue
-		current_indent = len(body) - len(stripped)
-		# Sibling def/class at parent's level (or shallower) → end of body.
-		sentinel_match = sentinel_re.match(body)
-		if sentinel_match and current_indent <= parent_indent:
-			break
-		if current_indent <= parent_indent:
-			# We've exited the parent function's block.
-			break
-		if call_re.search(body):
-			return i + 1  # convert 0-based index back to lineno
-
-	return None
-
-
-def _retarget_phase1_callsites_to_drilldown_leaf(
-	findings: list[dict],
-	file_cache: dict | None = None,
-) -> None:
-	"""Re-aim phase-1 finding callsites at the **call site** of the
-	deepest user-code frame in their drill-down chain, in place.
-
-	A Slow Hot Path / Hook Bottleneck / Repeated Hot Frame finding's
-	default callsite is its **wrapper's** entry frame (the ``def`` line
-	of the function at which the slow path was first flagged, e.g.
-	``looped_validate:6``). The drill-down chain already walks down to
-	the deepest user-code frame (e.g. ``_check_user_exists:18``), but
-	even that frame's ``def`` line is a function header rather than the
-	expensive call. The reader's eye lands on the most actionable info
-	when the snippet shows the **call expression** for the deepest
-	leaf — typically the line inside the **parent** of the deepest
-	frame that invokes it. For the user's case the chain is
-	``looped_validate → _run_validations → _check_user_exists`` and
-	the relevant line is ``    _check_user_exists(doc)`` at
-	``_run_validations:13``.
-
-	Phase-1 only — no phase-2 dependency. The drill-down chain is
-	gated on ``action_ref + tree + walk produces ≥1 step`` and is
-	already populated by ``_attach_drilldown_chains`` (called earlier
-	in the render flow). Most Slow Hot Path findings carry one;
-	findings without a chain are left as-is.
-
-	Hot Line / Function Not Invoked findings are phase-2 native — their
-	callsite is already at the leaf — so they're skipped. SQL "red
-	flag" findings whose callsite is a representative one (a query hot
-	spot) are skipped too: their snippet should keep pointing at the
-	query line.
-
-	The original wrapper location is preserved on the new callsite as
-	``original_wrapper`` (``{filename, lineno, function}``) so the
-	template can render a "Time entered through …" caption beneath
-	the relocated callsite header.
-
-	Fallback chain when the call site can't be located:
-	  1. The parent's body is parsed via AST; if a matching ``Call`` is
-	     found, anchor on its lineno.
-	  2. A regex fallback scans the parent's body for ``<callee>(``.
-	  3. If neither finds the call line, fall back to the leaf's own
-	     ``def`` lineno (previous behavior — still better than the
-	     wrapper's def line).
-	"""
-	if not findings:
-		return
-
-	for finding in findings:
-		ftype = finding.get("finding_type") or ""
-		if ftype in {"Hot Line", "Function Not Invoked"}:
-			continue
-		detail = finding.get("technical_detail")
-		if not isinstance(detail, dict):
-			continue
-		chain = detail.get("drilldown_chain") or []
-		if not chain:
-			continue
-		callsite = detail.get("callsite") or {}
-		# Skip representative SQL callsites — they're query hot spots,
-		# not function frames; re-targeting them would lose the query
-		# line they exist to point at.
-		if callsite.get("is_representative"):
-			continue
-		fname = callsite.get("filename") or ""
-		fn_name = callsite.get("function") or ""
-		if not fn_name:
-			continue
-		leaf = chain[-1] if isinstance(chain[-1], dict) else None
-		if not leaf:
-			continue
-		leaf_function = leaf.get("function") or ""
-		leaf_lineno = leaf.get("lineno")
-		leaf_filename = leaf.get("filename") or ""
-		if not leaf_function or leaf_lineno is None or not leaf_filename:
-			continue
-		# Single-frame chain (or leaf == wrapper) — nothing to relocate.
-		if leaf_function == fn_name:
-			continue
-
-		# Determine the leaf's caller in the chain: chain[-2] when the
-		# chain has an intermediate, else the wrapper (the finding's
-		# original callsite).
-		if len(chain) >= 2 and isinstance(chain[-2], dict):
-			parent_filename = chain[-2].get("filename") or ""
-			parent_lineno = int(chain[-2].get("lineno") or 0)
-			parent_function = chain[-2].get("function") or ""
-		else:
-			parent_filename = fname
-			parent_lineno = int(callsite.get("lineno") or 0)
-			parent_function = fn_name
-
-		call_lineno = _find_call_line_in_function_body(
-			parent_filename, parent_lineno, leaf_function,
-			file_cache=file_cache,
-		)
-
-		# Choose the anchor based on what we found:
-		#   - call site in the parent → ``(parent_filename, call_lineno, parent_function)``
-		#   - fallback: leaf's own def line ``(leaf_filename, leaf_lineno, leaf_function)``
-		if call_lineno is not None:
-			anchor_filename = parent_filename
-			anchor_lineno = call_lineno
-			anchor_function = parent_function
-		else:
-			anchor_filename = leaf_filename
-			anchor_lineno = leaf_lineno
-			anchor_function = leaf_function
-
-		snippet = _read_source_snippet(
-			anchor_filename, anchor_lineno, cache=file_cache,
-		)
-
-		display_filename = (
-			_bench_relative_display(anchor_filename)
-			if anchor_filename.startswith("/")
-			else anchor_filename
-		)
-		new_callsite = {
-			"filename": display_filename,
-			"_abs": (
-				anchor_filename
-				if anchor_filename.startswith("/")
-				else callsite.get("_abs")
-			),
-			"lineno": anchor_lineno,
-			"function": anchor_function,
-			"source_snippet": snippet,
-			"original_wrapper": {
-				"filename": fname,
-				"lineno": callsite.get("lineno"),
-				"function": fn_name,
-			},
-			# v0.7.x: the Phase-2 cross-link callout's lookup keys on
-			# ``(basename, function)`` — preserve the leaf's identity
-			# here so the template can target the leaf's phase-2 hot
-			# line even when the smoking-gun callsite was retargeted
-			# to the parent's call site (different function name).
-			"phase2_lookup_filename": leaf_filename,
-			"phase2_lookup_function": leaf_function,
-		}
-		detail["callsite"] = new_callsite
-
+# v0.12.26+: _find_call_line_in_function_body +
+# _retarget_phase1_callsites_to_drilldown_leaf moved to
+# optimus/renderer/finding_enrichment.py (phase 3). Re-imported
+# at the top of this module so call sites resolve unchanged.
 
 # v0.12.16+: _GROUPING_SEVERITY_RANK / _root_cause_key /
 # _group_findings_by_root_cause moved to
@@ -1577,408 +1319,11 @@ def build_background_jobs(actions, recordings_by_uuid, findings=None, tracked_jo
 # this module so call sites resolve unchanged.
 
 
-def _finding_to_dict(child: Any, file_cache: dict | None = None) -> dict:
-	"""Flatten a Optimus Finding child row, parsing the JSON detail blob.
-
-	v0.6.0 Round 2: synthesize a unified ``callsite`` shape for findings
-	that store their location at the top level (call_tree's Slow Hot Path
-	/ Hook Bottleneck / Repeated Hot Frame use top-level
-	``filename``/``lineno``; line_profile's Hot Line / Function Not
-	Invoked use top-level ``file``/``lineno``). Without this, the
-	smoking-gun block in finding_card never renders for those types.
-
-	Also lazily attach a ±1 source snippet to the callsite when one isn't
-	already persisted — covers (a) sessions analyzed before the
-	analyze-time enrichment shipped, (b) the synthesized callsites above.
-	The optional ``file_cache`` is shared across all findings in the same
-	render so a cluster of findings in one source file reads the file
-	once.
-	"""
-	try:
-		detail = json.loads(child.technical_detail_json or "{}")
-	except Exception:
-		detail = {}
-
-	# v0.6.0 Round 2: synthesize callsite from legacy top-level shape
-	# when the analyzer didn't wrap it. ``filename`` is the canonical
-	# key; ``file`` is the line-profile alias. Both are accepted.
-	if not detail.get("callsite"):
-		fname = detail.get("filename") or detail.get("file")
-		lineno = detail.get("lineno")
-		if fname and lineno is not None:
-			detail["callsite"] = {
-				"filename": fname,
-				"lineno": lineno,
-				"function": detail.get("function") or "",
-			}
-
-	# v0.5.3: normalize callsite shape so downstream code (app
-	# bucketing, template) can assume dict.
-	if "callsite" in detail:
-		detail["callsite"] = _normalize_callsite(detail.get("callsite"))
-
-	# v0.6.x: findings that name a function but carry no file:line — resolve
-	# them so the smoking-gun block can render. Repeated Hot Frame stores a
-	# redacted ``"path::func"`` key in ``function``; Function Not Invoked
-	# (phase 2) stores a ``dotted_path``. Best-effort, render-time.
-	if not (detail.get("callsite") or {}).get("lineno"):
-		_ftype = child.finding_type or ""
-		if _ftype == "Repeated Hot Frame" and detail.get("function"):
-			_cs = _resolve_frame_key_to_callsite(detail["function"], cache=file_cache)
-			if _cs:
-				detail["callsite"] = _cs
-		elif _ftype == "Function Not Invoked" and detail.get("dotted_path"):
-			_resolved = _resolve_dotted_to_code(
-				detail["dotted_path"], file_cache=file_cache,
-			)
-			if _resolved:
-				_abs, _ln, _name = _resolved
-				detail["callsite"] = {
-					"filename": _bench_relative_display(_abs),
-					"_abs": _abs,
-					"lineno": _ln,
-					"function": _name or str(detail["dotted_path"]).rsplit(".", 1)[-1],
-					"source_snippet": _read_source_snippet(_abs, _ln, cache=file_cache),
-				}
-
-	# v0.6.0 Round 2 / v0.7.x: a Hot Line finding carries the exact profiled
-	# line in ``line_content``. Show it WITH ±2 context lines (read from the
-	# file, like call-tree findings) instead of a lone row — but keep the
-	# profiled text authoritative for the hot line itself, so a file that
-	# drifted since the run can't misrepresent what was actually profiled.
-	# Falls back to the single stored line when the file can't be read at
-	# render (offline / regenerated without the source) or the line is gone.
-	callsite = detail.get("callsite") or {}
-	if (
-		callsite
-		and not callsite.get("source_snippet")
-		and detail.get("line_content")
-		and callsite.get("lineno") is not None
-	):
-		hot_ln = callsite["lineno"]
-		window = _read_source_snippet(
-			callsite.get("filename"), hot_ln, cache=file_cache,
-		)
-		if window and any(r.get("lineno") == hot_ln for r in window):
-			for r in window:
-				if r.get("lineno") == hot_ln:
-					r["content"] = detail["line_content"]  # profiled text wins
-			callsite["source_snippet"] = window
-		else:
-			callsite["source_snippet"] = [{
-				"lineno": hot_ln,
-				"content": detail["line_content"],
-			}]
-		detail["callsite"] = callsite
-
-	# v0.6.0 Round 2: lazy snippet read at render time when nothing has
-	# been attached yet (covers older sessions + synthesized callsites
-	# without a line_content shortcut).
-	if (
-		callsite
-		and callsite.get("filename")
-		and callsite.get("lineno") is not None
-		and not callsite.get("source_snippet")
-	):
-		snippet = _read_source_snippet(
-			callsite["filename"], callsite["lineno"], cache=file_cache,
-		)
-		if snippet:
-			callsite["source_snippet"] = snippet
-			detail["callsite"] = callsite
-
-	# v0.7.x+: Server Script findings carry the synthetic ``<serverscript>:
-	# <name>`` filename. Set ``_abs`` to the Desk URL (``/app/server-script/
-	# <actual-name>``) and tag ``_link_kind = "desk"`` so the template renders
-	# a click-to-Desk link instead of building a broken vscode://file URL
-	# over the synthetic path. Bare ``<serverscript>`` (no name) falls
-	# through with no link — same fallback as any other unresolvable callsite.
-	if callsite and callsite.get("filename"):
-		_fn = str(callsite["filename"])
-		if _fn.startswith("<serverscript") or _fn.startswith("<server-script"):
-			from optimus.server_script_source import desk_url, extract_script_name
-
-			_scrubbed = extract_script_name(_fn)
-			if _scrubbed:
-				callsite["_abs"] = desk_url(_scrubbed, cache=file_cache)
-				callsite["_link_kind"] = "desk"
-				detail["callsite"] = callsite
-
-	# v0.6.0: AI-suggested fix (on-demand; empty until generated). Stored as
-	# JSON on the child row by api.suggest_fix. Convert the Markdown body to
-	# sanitized HTML here so the template can `| safe` it (mirrors the
-	# notes_html pattern). A parse/convert failure falls back to an escaped
-	# <pre> so raw text is never rendered verbatim.
-	llm_fix = None
-	try:
-		raw_llm = json.loads(getattr(child, "llm_fix_json", None) or "{}")
-	except Exception:
-		raw_llm = {}
-	if isinstance(raw_llm, dict) and (raw_llm.get("suggestion") or "").strip():
-		llm_fix = {
-			"suggestion_html": _markdown_to_safe_html(raw_llm.get("suggestion")),
-			"model": raw_llm.get("model") or "",
-			"provider": raw_llm.get("provider") or "",
-			"generated_at": raw_llm.get("generated_at") or "",
-			# Older rows (pre this field) → assume the AI had context, so we
-			# don't slap a "directional only" caveat on suggestions that were
-			# in fact grounded in source.
-			"source_available": raw_llm.get("source_available", True),
-		}
-
-	return {
-		"finding_type": child.finding_type or "",
-		"severity": child.severity or "Low",
-		"title": child.title or "",
-		"customer_description": child.customer_description or "",
-		"estimated_impact_ms": child.estimated_impact_ms or 0,
-		"affected_count": child.affected_count or 0,
-		"action_ref": child.action_ref or "",
-		"technical_detail": detail,
-		"llm_fix": llm_fix,
-	}
-
-
-# v0.6.x: SQL "red flag" findings (Missing Index, Full Table Scan, Filesort,
-# Temporary Table, Low Filter Ratio) are keyed by (finding_type, table) and
-# carry no callsite — the offending query is issued from many places. At
-# render time we still have the recordings, so we pick a *representative*
-# callsite: the hottest user-app frame among the calls whose normalized query
-# matches the finding's. Best-effort — surfaced as "Most-called from:" with a
-# "representative callsite" note in the template.
-_SQL_REDFLAG_FINDING_TYPES = frozenset({
-	"Missing Index", "Full Table Scan", "Filesort", "Temporary Table",
-	"Low Filter Ratio",
-})
-
-
-# v0.12.19+: _find_node_in_tree / _walk_drilldown_chain /
-# _attach_drilldown_chains moved to optimus/renderer/finding_enrichment.py.
-# Re-imported at the top of this module so call sites resolve unchanged.
-
-
-def _attach_representative_callsites(findings, recordings, *, file_cache: dict | None = None) -> None:
-	"""Attach a representative ``callsite`` (+ ``is_representative``) to SQL
-	red-flag findings by matching their normalized query against the recording
-	calls and picking the hottest user-app frame. Mutates ``findings`` (the
-	``_finding_to_dict`` output dicts) in place. No-op when there are no such
-	findings, no recordings, or nothing matches — those cards just render
-	without the block.
-	"""
-	if not findings or not recordings:
-		return
-	wanted: list[dict] = []
-	for f in findings:
-		if (f.get("finding_type") or "") not in _SQL_REDFLAG_FINDING_TYPES:
-			continue
-		detail = f.get("technical_detail") or {}
-		if (detail.get("callsite") or {}).get("lineno"):
-			continue  # already has one
-		nq = (detail.get("normalized_query") or "").strip()
-		if not nq:
-			continue
-		wanted.append({
-			"finding": f,
-			"nq": nq,
-			"table": (detail.get("table") or "").strip(),
-			"tally": {},  # (filename, lineno, function) → weight
-		})
-	if not wanted:
-		return
-
-	try:
-		from optimus.analyzers.base import walk_callsite
-	except Exception:
-		return
-
-	for rec in recordings:
-		if not isinstance(rec, dict):
-			continue
-		for call in rec.get("calls") or []:
-			if not isinstance(call, dict):
-				continue
-			cnq = (call.get("normalized_query") or "").strip()
-			if not cnq:
-				continue
-			cquery = call.get("query") or ""
-			for w in wanted:
-				# Equality or prefix either way (survives truncation), plus
-				# the table name must appear in the raw query.
-				if not (cnq == w["nq"] or cnq.startswith(w["nq"]) or w["nq"].startswith(cnq)):
-					continue
-				if w["table"] and w["table"] not in cquery:
-					continue
-				frame = walk_callsite(call.get("stack"))
-				if not frame or not frame.get("filename") or frame.get("lineno") is None:
-					continue
-				k = (frame.get("filename"), frame.get("lineno"), frame.get("function") or "")
-				w["tally"][k] = w["tally"].get(k, 0) + (call.get("duration") or 0) + 1
-
-	for w in wanted:
-		if not w["tally"]:
-			continue
-		(filename, lineno, function), _weight = max(w["tally"].items(), key=lambda kv: kv[1])
-		abs_path = _resolve_source_path(filename)
-		w["finding"]["technical_detail"]["callsite"] = {
-			"filename": _bench_relative_display(abs_path) if abs_path else filename,
-			"_abs": abs_path,
-			"lineno": lineno,
-			"function": function,
-			"source_snippet": _read_source_snippet(abs_path or filename, lineno, cache=file_cache),
-			"is_representative": True,
-		}
-
-
-# v0.12.10+: doc-event lifecycle binding + per-DocType breakdown moved to
-# optimus/renderer/doc_event_renderer.py. The functions (_extract_target_doc,
-# _attach_action_context, _build_doc_event_breakdown, _doctype_from_controller_path,
-# _module_from_filename, _doc_event_hook_index, _finding_hook_events,
-# _finding_lifecycle_bindings, _build_doc_event_hook_index) + constants
-# (_LIFECYCLE_EVENTS, _KIND_*, _SEVERITY_RANK) are re-imported at the top of
-# this module so call sites resolve unchanged.
-
-
-def _markdown_to_safe_html(text) -> str:
-	"""Render Markdown → sanitized HTML for embedding in the report.
-
-	Mirrors the notes sanitization path (``frappe.utils.markdown`` +
-	``sanitize_html(..., always_sanitize=True)``). On ANY failure — Frappe
-	not importable, markdown/bleach hiccup — falls back to an HTML-escaped
-	``<pre>`` block so the report NEVER renders un-sanitized model output.
-
-	After sanitizing, fenced ``diff`` code blocks (which the AI-fix prompt
-	asks the model to use for before/after) get per-line ``dh-add`` /
-	``dh-del`` / ``dh-meta`` span wrappers so the report CSS can colour them
-	like a real diff. We only add ``<span>`` wrappers around already-escaped
-	text — nothing that could re-introduce unsafe markup.
-	"""
-	raw = "" if text is None else str(text)
-	try:
-		from frappe.utils import markdown as _md
-		from frappe.utils.html_utils import sanitize_html
-		return _highlight_diff_html(sanitize_html(_md(raw), always_sanitize=True))
-	except Exception:
-		import html as _html
-		return _highlight_diff_html(
-			'<pre style="white-space:pre-wrap;">' + _html.escape(raw) + "</pre>"
-		)
-
-
-# v0.10.0+: _PRE_BLOCK_RE, _looks_like_diff, _diff_line_class, and
-# _highlight_diff_html moved to optimus/renderer/syntax.py with the rest
-# of the Pygments + diff-highlighting helpers. The imports at the top of
-# this file re-introduce ``_highlight_diff_html`` under the same name.
-
-
-# v0.10.0+: _SNIPPET_TRUNCATE_CHARS, _path_within_bench, _resolve_source_path,
-# _read_source_snippet, _read_source_window all moved to
-# optimus/renderer/source.py — imported at the top of this file.
-# (Dead code that was moved to optimus/renderer/source.py has been removed here.
-# _resolve_source_path / _read_source_snippet / _read_source_window / _path_within_bench
-# / _SNIPPET_TRUNCATE_CHARS / _BoundedFileCache / _FILE_CACHE_MAX_ENTRIES now live in
-# optimus/renderer/source.py and are re-imported at the top of this file.)
-
-
-def _read_function_body_snippet(
-	filename: str,
-	def_lineno,
-	*,
-	cache: dict | None = None,
-	max_lines: int = 40,
-) -> list[dict] | None:
-	"""v0.7.x: read a whole function body — from its ``def`` line to the end
-	of the function — as ``[{lineno, content}]`` (same shape as
-	``_read_source_snippet``). Used for self-time hot-path findings with no
-	deeper user-code frame: Phase-1 sampling can't pinpoint the hot line, but
-	the function is the relevant unit, so show all of it rather than a ±2-line
-	peek. The function ends at the first non-blank line indented at or below
-	the ``def``'s indentation (next def/class/module statement); capped at
-	``max_lines``. Returns ``None`` when the file isn't readable / lineno is
-	out of range."""
-	try:
-		ln = int(def_lineno)
-	except (TypeError, ValueError):
-		return None
-	if ln <= 0 or not filename:
-		return None
-
-	if cache is not None and filename in cache:
-		lines = cache[filename]
-	else:
-		resolved = _resolve_source_path(filename)
-		if isinstance(resolved, tuple) and resolved[0] == "server_script":
-			from optimus.server_script_source import get_server_script_lines
-
-			lines = get_server_script_lines(resolved[1], cache=cache)
-		else:
-			try:
-				with open(resolved, encoding="utf-8") as fh:
-					lines = fh.read().splitlines()
-			except Exception:
-				lines = None
-		if cache is not None:
-			cache[filename] = lines
-
-	if not lines or ln > len(lines):
-		return None
-
-	limit = _SNIPPET_TRUNCATE_CHARS
-
-	def _row(idx: int) -> dict:
-		content = lines[idx - 1]
-		if len(content) > limit:
-			content = content[:limit] + "..."
-		return {"lineno": idx, "content": content}
-
-	def_line = lines[ln - 1]
-	def_indent = len(def_line) - len(def_line.lstrip())
-
-	out = [_row(ln)]
-	n = ln + 1
-	while n <= len(lines) and len(out) < max_lines:
-		raw = lines[n - 1]
-		if raw.strip():
-			indent = len(raw) - len(raw.lstrip())
-			if indent <= def_indent:
-				break  # dedented to def level — function (and decorators) ended
-		out.append(_row(n))
-		n += 1
-	return out or None
-
-
-def _expand_self_time_snippets(findings, *, file_cache: dict | None = None) -> None:
-	"""v0.7.x: for self-time hot-path findings with no deeper user-code frame
-	(empty ``drilldown_chain``), narrow the smoking-gun snippet to the function's
-	signature line and flag it ``self_time_no_pinpoint``. Phase-1 sampling can't
-	pinpoint a single hot line inside the function, so dumping the whole body
-	(highlighting only the def) was a misleading wall of code — the card now
-	shows just the def + a note pointing the developer at a Line-Level Drilldown
-	on the function (which CAN give per-line timing).
-
-	Runs AFTER ``_attach_drilldown_chains`` (which populates ``drilldown_chain``)
-	and mutates findings in place. Best-effort: only the empty-list case (a
-	deeper chain means the ±2 window + chain is enough; a missing key means the
-	chain was never computed)."""
-	for finding in findings or []:
-		if (finding.get("finding_type") or "") != "Slow Hot Path":
-			continue
-		detail = finding.get("technical_detail") or {}
-		# Only when a chain was attempted AND came back empty (no deeper frame).
-		if detail.get("drilldown_chain") != []:
-			continue
-		callsite = detail.get("callsite") or {}
-		fn = callsite.get("filename")
-		ln = callsite.get("lineno")
-		if not fn or ln is None:
-			continue
-		# Flag drives the "no single hot line — run a Line-Level Drilldown" note.
-		callsite["self_time_no_pinpoint"] = True
-		body = _read_function_body_snippet(fn, ln, cache=file_cache)
-		if body:
-			callsite["source_snippet"] = body[:1]  # signature/def line only
-		detail["callsite"] = callsite
-		finding["technical_detail"] = detail
-
+# v0.12.26+: _finding_to_dict + _SQL_REDFLAG_FINDING_TYPES +
+# _attach_representative_callsites + _markdown_to_safe_html +
+# _read_function_body_snippet + _expand_self_time_snippets moved
+# to optimus/renderer/finding_enrichment.py (phase 3). Re-imported
+# at the top of this module so call sites resolve unchanged.
 
 # v0.12.23+: source-resolution helpers (_action_dotted_entry,
 # _skip_decorators_to_def, _resolve_dotted_to_code, _bench_relative_display,
